Bug fix: integer instead of real division in animating appear/disappear logic

+++ source/Gamee/Actor.cpp
@@ -86,7 +86,7 @@ int Actor::animateAppear(QObject* object, int duration, int initialDelay, qreal 
 	// Do not change initial value on delay
 	animation->setKeyValueAt(0.0, fromOpacity);
 	if ( totalDuration > 0 )
-		animation->setKeyValueAt(initialDelay / totalDuration, fromOpacity);
+		animation->setKeyValueAt(qreal(initialDelay) / totalDuration, fromOpacity);
 	animation->setKeyValueAt(1.0, toOpacity);
 
 	animation->start();
